+++ Picmory/Models/Repositorys/SQLFollowerRepository.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using Picmory.Models.RequestResultModels;
+using Picmory.Util;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -15,36 +16,49 @@ namespace Picmory.Models.Repositorys
         }
 
 
-        public bool AnswerNewFollower(bool accept, User follower, User followed)
+        public Success AnswerNewFollower(bool accept, User follower, User followed)
         {
             var follow = context.Followers.FirstOrDefault(item => item.Follower == follower && item.Followed == followed);
             if (follow != null)
             {
+                if (follow.Accepted != null) { return Success.FailedByAlreadyAnswered; }
                 follow.Accepted = accept;
                 context.SaveChanges();
-                return true;
+                return Success.Successfull;
             }
-            return false;
+            return Success.FailedByNotRequested;
         }
 
-        public Followers AskNewFollower(User follower, User followed)
+        public Success AskNewFollower(User follower, User followed)
         {
-            Followers follow = new Followers(followed, follower, null);
-            context.Followers.Add(follow);
-            context.SaveChanges();
-            return follow;
+            Followers requested = context.Followers
+                    .Include(a => a.Follower)
+                    .Where(a => a.Follower == follower && a.Followed == followed)
+                    .FirstOrDefault();
+            if (requested == null)
+            {
+                Followers follow = new Followers(followed, follower, null);
+                context.Followers.Add(follow);
+                context.SaveChanges();
+                return Success.Successfull;
+            }
+            else if (requested.Accepted == null)
+                { return Success.FailedByAlreadyRequested; }
+            else
+                { return Success.FailedByAlreadyFollowed; }
         }
 
-        public bool DeleteFollower(User follower, User followed)
+        public Success DeleteFollower(User follower, User followed)
         {
             var follow = context.Followers.FirstOrDefault(item => item.Follower == follower && item.Followed == followed);
             if (follow != null)
             {
+                if (follow.Accepted == null) { return Success.FailedByNotAccepted; }
                 context.Followers.Remove(follow);
                 context.SaveChanges();
-                return true;
+                return Success.Successfull;
             }
-            return false;
+            return Success.FailedByNotExist;
         }
 
         public List<string> GetAllFollowers(User user)
